create abstract point dto
