+++ src/main/java/com/vsu/amm/medframe/dto/AbstractPointDto.java
@@ -1,0 +1,35 @@
+package com.vsu.amm.medframe.dto;
+
+class AbstractPointDto {
+
+    protected Long id;
+
+    protected int frequency;
+
+    protected int intensityLevel;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public int getFrequency() {
+        return frequency;
+    }
+
+    public void setFrequency(int frequency) {
+        this.frequency = frequency;
+    }
+
+    public int getIntensityLevel() {
+        return intensityLevel;
+    }
+
+    public void setIntensityLevel(int intensityLevel) {
+        this.intensityLevel = intensityLevel;
+    }
+
+}
