+++ src/main/java/com/vsu/amm/medframe/dto/TestDto.java
@@ -1,0 +1,6 @@
+package com.vsu.amm.medframe.dto;
+
+public class TestDto {
+
+    private Long id;
+}
