+++ src/main/java/com/vsu/amm/medframe/dto/TestPointDto.java
@@ -1,0 +1,4 @@
+package com.vsu.amm.medframe.dto;
+
+public class TestPointDto extends AbstractPointDto{
+}
